// chpqa-app/src/app/api-services/chpqa-api/generated/model/requestCreateManufactModelEngineUnit.ts
/**
 * WebApi
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface RequestCreateManufactModelEngineUnit { 
    idSubmission: string;
    manufacturerOther?: string | null;
    modelOther?: string | null;
    engineUnitOther?: string | null;
    totalPowerCapacityKw?: number | null;
    totalHeatCapacityKw?: number | null;
    fuelInputKw?: number | null;
    powerEfficiency?: number | null;
    maxHeatToPowerRatio?: number | null;
    maxHeatEfficiency?: number | null;
    maxOverallEfficiency?: number | null;
}